requirements.txt
pip~=23.2.1
wheel~=0.41.2
setuptools~=68.2.0
Django~=4.2
Jinja2~=3.1.2
pytz~=2023.3.post1
colorama~=0.4.6
click~=8.1.7
asgiref~=3.7.2
sqlparse~=0.4.4
djangorestframework~=3.14.0
attrs~=23.1.0
Pillow~=10.1.0
black~=23.12.0
platformdirs~=4.1.0
tomli~=2.0.1
packaging~=23.2
pathspec~=0.12.1
PyYAML~=6.0.1
uritemplate~=4.1.1
psycopg2==2.9.9
python-dotenv==1.0.0
drf-spectacular==0.27.0
django-debug-toolbar==4.2.0
djangorestframework-simplejwt==5.3.1


blinker==1.7.0
django-rest-framework==0.1.0
Flask==3.0.0
inflection==0.5.1
itsdangerous==2.1.2
jsonschema==4.20.0
legacy==0.1.7
MarkupSafe==2.1.3
modcall==0.1.0
mypy-extensions==1.0.0
namespace==0.1.4
PyJWT==2.8.0
referencing==0.32.0
rpds-py==0.13.2
tzdata==2023.3
Werkzeug==3.0.1
